Traducción
To secure an online account, change the password to a unique phrase or code that only you would know. When creating a password, don't use any words or phrases your partner could guess. Random combinations of words, such as "PurpleDishwasherPotato", are easily remembered and very difficult to hack. It's especially important that you don't re-use a password from another account. If you're changing a lot of a passwords, we strongly recommend using a password manager, which generates and stores passwords so you don't have to remember them. This makes it nearly impossible for a partner to guess your password. It's also much safer than writing passwords in a document, or on a piece of paper. Read our Protect Your Passwords defense strategy for more detailed information.

Strengthen